Create more controls in UI for managing Runs
For task DELFT3DGT-107. Startmodel button uses the earlier made
"runmodel" on the server side. Deletemodel uses the previous
deletemodel.

When the checkmark is pressed, an additional "deletefiles = true" is
passed in the POST data. "Start" is greyed out when the model is in
"processing" state. (need to add other states later). Messagebox shows
extra alert when the delete checkmark is checked.

Please note this issue depends on DELFT3DGT-109 at the webserver side.