Made changes based upon comments
Made changes based upon comments here:

https://issuetracker.deltares.nl/browse/DELFT3DGT-97?focusedCommentId=528801&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-528801

Note that properties of objects for the data exchange should be, and
will be always lower case. Treat it like filename. If everyone knows it
should be lower case, it makes the live of everyone better.

Also fixed some SCSS issues as the build server explains.